requirements.txt
aiogram==2.25.1
openai==0.27.6
Pillow==9.5.0
pydub==0.25.1
pytesseract==0.3.10
Unidecode==1.3.6
